cal: add landscape layout for yearly views
With the high availability of wide screens and high-resolution emulated
terminals and terminal window managers it's sometimes useful to change
the output of cal(1) yearly views to more stretched out ones
horizontally.

This patch adds a new -l/--landscape flag that stretches out the yearly
view to 3x4 and 4x3 months in Julian mode and 4x3 and 6x2 in regular
mode. The flag can be used up to two times.