agetty: dont use default speed for serial console
Serial console is used for boot messages and is pre-configured by kernel.
Previous versions of agetty didn't change its speed.
To keep the backward compatibility, detect the serial console and
don't change its speed settings if none are explicitly specified.

This treatment is similar to how the virtual consoles are handled.