hwclock: remove automatic ISA fallback
The old behavior when hwclock was unable to access
the rtc device was to fallback to direct ISA I/O.

This was sometimes allowing two processes to access
the Hardware Clock at the same time and hanging the
system. This is the reason the rtc device driver
only allows one process to open it.

Now hwclock defaults to using the rtc device and only
uses direct ISA when that option is given.

Signed-off-by: J William Piggott <elseifthen@gmx.com>