sulogin: Don't ask for password when it is locked/disabled
Some installations and distributions don't use a root account password for
security reasons and use sudo instead. In that case, asking for the password
makes no sense, and it is not even considered as valid as it's just "*" or "!".

In these cases, just start a root shell. As both sulogin and getting into
single user mode/emergency.target require root access or physical hardware
access anyway, this is not a privilege escalation.

Addresses: https://bugs.debian.org/326678

Signed-off-by: Martin Pitt <martin.pitt@ubuntu.com>
Reviewed-by: Andreas Henriksson <andreas@fatal.se>